Extract Delta's to a seperated module
